split code of client + clean abortion of not implemented udp
