+++ srp/src/main.rs
@@ -1,23 +1,24 @@
-use log::{error, info};
+use log::{info};
 use quinn::crypto::rustls::QuicClientConfig;
 use quinn::{ClientConfig, Connection, Endpoint};
-use shared::{ClientConfigRequest, ServerConfigResponse};
+use shared::{ClientConfigRequest, Protocol, ServerConfigResponse};
 use std::{
     net::{IpAddr, Ipv4Addr, SocketAddr},
     sync::Arc,
 };
-use tokio::{
-    io::AsyncWriteExt,
-    net::TcpStream,
-    signal,
-    task::JoinSet,
-};
+use tokio::signal;
 use tokio_util::sync::CancellationToken;
+
+use shared::MAX_MSG_SIZE;
 
 mod certificate_validation;
 use certificate_validation::SkipServerVerification;
 
-const MAX_MSG_SIZE: usize = 64 * 1024;
+mod proxy_tcp;
+use proxy_tcp::proxy_tcp_stream;
+
+mod proxy_udp;
+use proxy_udp::proxy_udp_stream;
 
 async fn configure_server(
     connection: Connection,
@@ -36,113 +37,37 @@ async fn configure_server(
         Ok(())
     } else {
         Err(anyhow::anyhow!(
-            response.error_message.unwrap_or_else(|| "unknown error".to_string())
+            response
+                .error_message
+                .unwrap_or_else(|| "unknown error".to_string())
         ))
     }
 }
 
-async fn handle_stream(
-    mut send: quinn::SendStream,
-    mut recv: quinn::RecvStream,
-    endpoint_addr: SocketAddr,
-) {
-    let result = async {
-        let tcp = TcpStream::connect(endpoint_addr).await?;
-        let (mut tcp_r, mut tcp_w) = tcp.into_split();
-
-        let quic_to_tcp = async {
-            tokio::io::copy(&mut recv, &mut tcp_w).await?;
-            tcp_w.shutdown().await
-        };
-
-        let tcp_to_quic = async {
-            tokio::io::copy(&mut tcp_r, &mut send).await?;
-            send.finish()?;
-            Ok::<_, std::io::Error>(())
-        };
-
-        tokio::try_join!(quic_to_tcp, tcp_to_quic)?;
-
-        Ok::<(), anyhow::Error>(())
-    }
-    .await;
-
-    if let Err(e) = result {
-        error!("[client stream] error: {e}");
-    }
-}
-
-async fn proxy_tcp_stream(
-    connection: Connection,
-    endpoint_addr: SocketAddr,
-    shutdown: CancellationToken,
-) -> anyhow::Result<()> {
-    let mut tasks = JoinSet::new();
-
-    loop {
-        tokio::select! {
-            _ = shutdown.cancelled() => {
-                info!("shutdown received, closing QUIC connection");
-                connection.close(0u32.into(), b"client shutdown".as_ref());
-                break;
-            }
-
-            accept = connection.accept_bi() => {
-                let (send, recv) = match accept {
-                    Ok(v) => v,
-                    Err(e) => {
-                        error!("accept_bi error: {e}");
-                        break;
-                    }
-                };
-
-                let addr = endpoint_addr;
-
-                tasks.spawn(async move {
-                    handle_stream(send, recv, addr).await;
-                });
-            }
-        }
-    }
-
-    while let Some(res) = tasks.join_next().await {
-        if let Err(e) = res {
-            error!("task join error: {e}");
-        }
-    }
-
-    Ok(())
-}
-
-async fn run_client(
+pub async fn run_client(
     server_socket: SocketAddr,
     endpoint_socket: SocketAddr,
     config_request: ClientConfigRequest,
     shutdown: CancellationToken,
 ) -> anyhow::Result<()> {
-    let mut endpoint =
-        Endpoint::client(SocketAddr::new(IpAddr::V4(Ipv4Addr::LOCALHOST), 0))?;
+    let mut endpoint = Endpoint::client(SocketAddr::new(IpAddr::V4(Ipv4Addr::LOCALHOST), 0))?;
 
     rustls::crypto::ring::default_provider()
         .install_default()
         .expect("crypto provider failed");
 
-    let client_config = ClientConfig::new(Arc::new(
-        QuicClientConfig::try_from(
-            rustls::ClientConfig::builder()
-                .dangerous()
-                .with_custom_certificate_verifier(SkipServerVerification::new())
-                .with_no_client_auth(),
-        )?,
-    ));
+    let client_config = ClientConfig::new(Arc::new(QuicClientConfig::try_from(
+        rustls::ClientConfig::builder()
+            .dangerous()
+            .with_custom_certificate_verifier(SkipServerVerification::new())
+            .with_no_client_auth(),
+    )?));
 
     endpoint.set_default_client_config(client_config);
 
-    info!("connecting to {server_socket}");
+    info!("connecting to {}", server_socket);
 
-    let connection = endpoint
-        .connect(server_socket, "localhost")?
-        .await?;
+    let connection = endpoint.connect(server_socket, "localhost")?.await?;
 
     info!("connected to {}", connection.remote_address());
 
@@ -156,7 +81,16 @@ async fn run_client(
             connection.close(0u32.into(), b"shutdown");
         }
 
-        res = proxy_tcp_stream(connection.clone(), endpoint_socket, shutdown.clone()) => {
+        res = async {
+            match config_request.protocol {
+                Protocol::Udp => {
+                    proxy_udp_stream(connection.clone(), endpoint_socket, shutdown.clone()).await
+                }
+                Protocol::Tcp => {
+                    proxy_tcp_stream(connection.clone(), endpoint_socket, shutdown.clone()).await
+                }
+            }
+        } => {
             res?;
         }
     }
@@ -197,13 +131,7 @@ async fn main() -> Result<(), Box<dyn std::error::Error + Send + Sync + 'static>
         shutdown_clone.cancel();
     });
 
-    run_client(
-        server_socket,
-        endpoint_socket,
-        config_request,
-        shutdown,
-    )
-    .await?;
+    run_client(server_socket, endpoint_socket, config_request, shutdown).await?;
 
     Ok(())
 }
